Moved NewMongoInfo and NewDialOpts to avoid import loop in replicaset
This reintroduces a small amount of duplication but is probably the
cleanest way to resolve this. It also means these functions are under
state again which makes more sense conceptually.
