+++ state/machine_test.go
@@ -448,8 +448,8 @@ func (s *MachineSuite) TestTag(c *gc.C) {
 }
 
 func (s *MachineSuite) TestSetMongoPassword(c *gc.C) {
-	info := coretesting.NewMongoInfo()
-	st, err := state.Open(info, coretesting.NewDialOpts(), state.Policy(nil))
+	info := testing.NewMongoInfo()
+	st, err := state.Open(info, testing.NewDialOpts(), state.Policy(nil))
 	c.Assert(err, jc.ErrorIsNil)
 	defer st.Close()
 	// Turn on fully-authenticated mode.
@@ -472,7 +472,7 @@ func (s *MachineSuite) TestSetMongoPassword(c *gc.C) {
 
 	// Check that we can log in with the correct password.
 	info.Password = "foo"
-	st1, err := state.Open(info, coretesting.NewDialOpts(), state.Policy(nil))
+	st1, err := state.Open(info, testing.NewDialOpts(), state.Policy(nil))
 	c.Assert(err, jc.ErrorIsNil)
 	defer st1.Close()
